Split mcp/tools.mts to satisfy max-lines lint rule
tools.mts (tool list + dispatcher), tool-handlers.mts (handler impls),
tool-coerce.mts (input coercion helpers) — each under 200 lines.
